backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.1
python-magic==0.4.27
PyPDF2==3.0.1
python-docx==0.8.11
httpx==0.25.1
groq==0.4.1
